<fix bug>

+++ GameLogic/src/main/java/nro/model/item/ItemOptionTemplate.java
@@ -1,4 +1,4 @@
 package nro.model.item;
 
-public record ItemOptionTemplate(int id, String name, int type) {
+public record ItemOptionTemplate(int id, String name, byte type) {
 }
